Refactor:리프레시 범위  수정
